+++ src/lib/node.cpp
@@ -2,45 +2,6 @@
 
 #include <ranges>
 
-void OldWorldImpl::update(UpdateContext& context) {
-    std::stable_sort(std::begin(m_nodes), std::end(m_nodes),
-                     [](const std::unique_ptr<Node>& a, const std::unique_ptr<Node>& b) {
-                         return a->getZIndex() < b->getZIndex();
-                     });
-
-    for (auto& node : m_nodes) {
-        node->update(context);
-    }
-};
-
-void OldWorldImpl::render(RenderContext& context) {
-    for (auto& node : m_nodes) {
-        node->render(context);
-    }
-}
-
-void OldWorldImpl::onMouseButtonEvent(MouseButtonEvent& event) {
-    for (auto it = m_nodes.rbegin(); it != m_nodes.rend(); it++) {
-        (*it)->onMouseButtonEvent(event);
-        if (event.isPropagationStopped()) {
-            break;
-        }
-    }
-};
-
-void OldWorldImpl::onMouseMotionEvent(MouseMotionEvent& event) {
-    for (auto it = m_nodes.rbegin(); it != m_nodes.rend(); it++) {
-        (*it)->onMouseMotionEvent(event);
-        if (event.isPropagationStopped()) {
-            break;
-        }
-    }
-};
-
-void OldWorldImpl::addNode(std::unique_ptr<Node> node) {
-    m_nodes.push_back(std::move(node));
-}
-
 void Node::initWithTextureRect(TextureRect textureRect) {
     setTextureRect(textureRect);
     setSize(textureRect.bounds.size);
